+++ modules/AWSPowerShell/Cmdlets/FSx/Basic/New-FSXVolumeFromBackup-Cmdlet.cs
@@ -128,9 +128,12 @@ namespace Amazon.PowerShell.Cmdlets.FSX
         #region Parameter OntapConfiguration_SecurityStyle
         /// <summary>
         /// <para>
-        /// <para>The security style for the volume. If a volume's security style is not specified,
-        /// it is automatically set to the root volume's security style. Specify one of the following
-        /// values:</para><ul><li><para><code>UNIX</code> if the file system is managed by a UNIX administrator, the majority
+        /// <para>Specifies the security style for the volume. If a volume's security style is not specified,
+        /// it is automatically set to the root volume's security style. The security style determines
+        /// the type of permissions that FSx for ONTAP uses to control data access. For more information,
+        /// see <a href="https://docs.aws.amazon.com/fsx/latest/ONTAPGuide/managing-volumes.html#volume-security-style">Volume
+        /// security style</a> in the <i>Amazon FSx for NetApp ONTAP User Guide</i>. Specify one
+        /// of the following values:</para><ul><li><para><code>UNIX</code> if the file system is managed by a UNIX administrator, the majority
         /// of users are NFS clients, and an application accessing the data uses a UNIX user as
         /// the service account. </para></li><li><para><code>NTFS</code> if the file system is managed by a Windows administrator, the majority
         /// of users are SMB clients, and an application accessing the data uses a Windows user